Revert "[artf625993]:Random behavior observed while configuring MAX WTX Count"
This reverts commit 04a7244f3a6be6fafeee472ee685b1fe8fb9adef.

+++ SN100x/src/nfa/include/nfa_api.h
@@ -639,7 +639,7 @@ typedef struct {
   /* Number of host in the whitelist of Terminal host */
 #if(NXP_EXTNS == TRUE)
   /* Maximum  time to wait for HCI response */
-  uint16_t max_wtx_count;
+  uint8_t max_wtx_count;
 #endif
   uint8_t num_whitelist_host;
   /* Whitelist of Terminal Host */
